Change `oneOf` to `anyOf` schema in owasp:api4:2019-string-restricted and owasp:api4:2019-string-limit (#43)
* Change `oneOf` to `anyOf` schema in owasp:api4:2019-string-restricted

The use of `oneOf` means that _only_ one of the schema constraints may be true,
but it is valid if two or more of the string schema constraint keywords
are present: `format`, `pattern`, `enum`, `const`.

* Change `oneOf` to `anyOf` schema in owasp:api4:2019-string-limit

Similar as with `2019-string-restricted`: the schema check should be
`anyOf` instead of `oneOf` in case a schema has more than one of the
keywords.